Modifica sapi_examples/cycles_counter_demo para formatear con stdioPrintf la salida. Elimina el sapi_examples/stdio_printf_and_buffer.

+++ modules/lpc4337_m4/sapi/src/sapi_cyclesCounter.c
@@ -101,7 +101,7 @@ void 		cyclesCounterReset	( void ){
  */
 float	cyclesCounterToUs	(uint32_t cycles){
 float valueInMicroSeconds = 0;
-	valueInMicroSeconds = cycles/(ClockSpeed/1000000);
+	valueInMicroSeconds = (float)cycles/(ClockSpeed/1000000);
 	return valueInMicroSeconds;
 }
 
@@ -115,7 +115,7 @@ float valueInMicroSeconds = 0;
 float	cyclesCounterToMs	(uint32_t cycles){
 	float valueInMilliSeconds = 0;
 
-	valueInMilliSeconds = cycles/(ClockSpeed/1000);
+	valueInMilliSeconds = (float)cycles/(ClockSpeed/1000);
 
 	return valueInMilliSeconds;
 }
